requirements.txt
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
matplotlib==3.7.2
requests==2.31.0